requirements.txt
transformers==4.38.2
torch
Pillow
